technical documentation in progress
